Move Manifestation to the manifestation module
