+++ rust-install/src/lib.rs
@@ -37,7 +37,7 @@ pub mod dist;
 mod errors;
 mod install;
 pub mod component;
-pub mod manifest;
+pub mod manifestation;
 pub mod download;
 
 // Mocks shared by tests
