+++ rust-install/src/manifestation.rs
