+++ rust-install/tests/dist.rs
@@ -19,7 +19,7 @@ use rust_install::dist::ToolchainDesc;
 use rust_install::download::DownloadCfg;
 use rust_install::utils;
 use rust_install::temp;
-use rust_install::manifest::{Manifestation, UpdateStatus, Changes};
+use rust_install::manifestation::{Manifestation, UpdateStatus, Changes};
 use rust_manifest::{Manifest, Component};
 use hyper::Url;
 use std::fs;
