+++ src/toolchain.rs
@@ -1,7 +1,7 @@
 use errors::*;
 use rust_install::{utils, dist, InstallPrefix, InstallType, InstallMethod};
 use rust_install::dist::ToolchainDesc;
-use rust_install::manifest::{Manifestation, Changes};
+use rust_install::manifestation::{Manifestation, Changes};
 use rust_manifest::Component;
 use config::Cfg;
